% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_log_entry.R
\name{create_log_entry}
\alias{create_log_entry}
\title{Create log entry}
\usage{
create_log_entry(event_name = NULL, event_body = NULL,
  event_type = "EVENT", event_status = "DONE", event_params = NULL,
  event_counter = -1, separator = "|")
}
\description{
\code{create_log_entry} creates a standardized log entry.
}
\details{
Function \code{create_log_entry} takes as arguments
different information about a single event and
creates standardized log entry header and body
used to be displayed in R console, browser JavaScript console,
or as a one-line log record in the log file
(after additional conversion from multi-line output).
}
