% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug_object.R
\name{debug_objects}
\alias{debug_objects}
\title{Copying objects to global environment}
\usage{
debug_objects(...)
}
\arguments{
\item{...}{Named objects to be copy to the global environment.
If there is only one unnamed object, its name in the global environment
will be the same as the name of the object passed to \code{...}.}
}
\description{
With \code{debug_objects} you can copy an object to the global environment
for further debugging or developing.
}
\examples{
\dontrun{

if (interactive()) {

  shiny::shinyApp(
    ui = shiny::fluidPage(log_init()),
    server = function(input, output) {
      set_logging()
      debug_objects(mtcars)
      debug_objects(df1 = head(mtcars), df2 = head(iris))
    }
  )

}
}
}
