% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_init.R
\name{log_init}
\alias{log_init}
\title{Initialize logging to JavaScript console}
\usage{
log_init()
}
\value{
A \code{tagList} with \code{script} tag inside \code{head} tag.
}
\description{
\code{log_init} should be put into the \code{shiny} \code{ui}
to initialize JavaScript code that enables logging to JavaScript console
in an Internet browser.
}
\examples{
\dontrun{

if (interactive()) {
  shiny::shinyApp(
    ui = shiny::fluidPage(log_init()),
    server = function(input, output) {
      set_logging(js_console = TRUE)
      log_event("See browser JavaScript console (CTRL + SHIFT + I)")
    }
  )
}
}
}
