% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_to_database.R
\name{log_to_database}
\alias{log_to_database}
\title{Logging to database}
\usage{
log_to_database(event_counter, event_type, event_name, event_status,
  event_params, event_body, event_timestamp, session_id)
}
\arguments{
\item{event_counter}{An integer.
The counter of events.
Default is \code{-1}.}

\item{event_type}{A character string.
The type of an event.
Default is \code{"EVENT"}.}

\item{event_name}{A character string or \code{NULL}.
The name of the event.
Default is \code{NULL}.}

\item{event_status}{A character string.
The status of an event.
Default is \code{"DONE"}.}

\item{event_params}{A list or \code{NULL}.
List of event-specific parameters.
Default is \code{NULL}.}

\item{event_body}{A character string or \code{NULL}.
The body of the entry in the eventlog.
Default is \code{NULL}.}

\item{event_timestamp}{A POSIXct or object convertable to POSIXct.}

\item{session_id}{A character string.}
}
\description{
Logging to database
}
\concept{low-level logging functions}
