% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_demo.R
\name{run_demo}
\alias{run_demo}
\alias{run_demo_dashboard}
\title{Run demo shiny app}
\usage{
run_demo(in_background = FALSE)

run_demo_dashboard()
}
\arguments{
\item{in_background}{A logical.
If \code{TRUE} the demo shiny app is run in the background on port 5555.
Default is \code{FALSE}.}
}
\description{
\code{run_demo} runs demo shiny app which logs different types of events.
\code{run_demo_dashboard} runs demo shiny dashboard that allows
for interactive analysis of events from demo app.
The demo app can be also run in background and events fired in the app
can be seen immediately in the demo dashboard.
}
\section{Functions}{
\itemize{
\item \code{run_demo}: Run demo shiny app

\item \code{run_demo_dashboard}: Run demo shiny dashboard
}}

\examples{
\donttest{
if (interactive()) {
  run_demo(in_background = TRUE)
  run_demo_dashboard()
}
}
}
