% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_logging.R
\name{set_logging}
\alias{set_logging}
\title{Setting events logging}
\usage{
set_logging(r_console = TRUE, js_console = TRUE, file = FALSE,
  database = FALSE, ...)
}
\arguments{
\item{r_console}{A logical. Should events be logged into R console?
Default is \code{TRUE}.}

\item{js_console}{A logical. Should events be logged into browser
JavaScript console? Default is \code{TRUE}.}

\item{file}{A logical or a character string.
Should events be logged to a file?
Default is \code{FALSE}.
If \code{TRUE} the default eventlog filename is \code{"events.log"}.
If character string, path and name of the filelog.}

\item{database}{A logical or a character string.
Should events be logged into a database?
  Default is \code{FALSE}.
  If \code{TRUE} or \code{"mongoDB"} the connection URL to the database
  will be read from the first line of a text file named \code{".db_url"}.}

\item{...}{a set of named objects
 (usually of type character, numeric, or date)
to be logged as parameters common to all events.}
}
\description{
\code{set_logging} should be used at the beginning of
the shiny server function to define where the logging should be done.
Events can be sent to R console, browser JavaScript console,
a eventlog file, or a database (or any combination of these).
By default logging is done to the R console and JavaScript console.
\code{set_logging} also sets event counter to 1 and can be used
to define global event parameters - named objects passed to \code{...}
that will be evaluated and added to lists of parameters of all events.
}
\details{
\code{set_logging} also assigns two new environments to the parent frame:
\code{log_settings_global} for storing global event parameters
and \code{log_event_register} for storing information about multiple
instances of the same event (see \code{\link{log_started}}).
If \code{database = TRUE} additional database connection object
named \code{log_db} is assigned to the parent frame.
}
\examples{
\dontrun{
if (interactive()) {

  shiny::shinyApp(
    ui = shiny::fluidPage(log_init()),
    server = function(input, output) {
      set_logging(r_console = TRUE, js_console = FALSE,
                  "param_1" = 1,
                  "param_2" = "A")
      log_event("Event with global param")
    }
  )
}
}
}
