% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyEventLogger.R
\docType{package}
\name{shinyEventLogger}
\alias{shinyEventLogger}
\alias{shinyEventLogger-package}
\title{shinyEventLogger: Logging Events in Complex Shiny Apps}
\description{
Logging framework dedicated for complex shiny apps. 
    Different types of events can be logged (character string, 
    value of a variable, multi-line output of a function,
    result of a unit test, custom error, warning, or diagnostic message).
    Each event can be logged with a list of parameters that are event-specific, 
    common for a group of events, or common for all app events.
    Logging can be done simultaneously to R console, 
    browser JavaScript console, a file log, and a database (currently MongoDB).
    Log data can be further analyzed with the help of process-mining techniques 
    from 'bupaR' package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kalimu/shinyEventLogger#readme}
  \item \url{https://kalimu.github.io/project/shinyeventlogger/}
  \item Report bugs at \url{https://github.com/kalimu/shinyEventLogger/issues}
}

}
\author{
Kamil Wais \email{kamil.wais@gmail.com}
}
