% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyEventLogger.R
\docType{package}
\name{shinyEventLogger}
\alias{shinyEventLogger}
\alias{shinyEventLogger-package}
\title{shinyEventLogger: Logging and Analysing Events in Shiny Apps}
\description{
Logging dedicated to Shiny apps. You can log different types
    of events: character string, current value of a variable, 
    multiline output of a function (eg. str(data.frame) or head(data.frame)), 
    result of a unit test, errors, warnings, diagnostic messages, etc.
    Logging can be done simultaneously to R console, 
    JavaScript console in a browser, or to a file log.
    Log data can be analysed with the help of dedicated functions 
    and process-mining techniques.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kalimu/shinyEventLogger}
  \item Report bugs at \url{https://github.com/kalimu/shinyEventLogger/issues}
}

}
\author{
Kamil Wais \email{kamil.wais@gmail.com}
}
