% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyEventLogger.R
\docType{package}
\name{shinyEventLogger}
\alias{shinyEventLogger}
\alias{shinyEventLogger-package}
\title{shinyEventLogger: Logging and Analyzing Events in Complex Shiny Apps}
\description{
Logging dedicated to Shiny apps. You can log different types
    of events, for example: character string, current value of a variable, 
    multi-line output of a function (e.g. str(data.frame) or head(data.frame)), 
    result of a unit test (using testthat), 
    custom errors, warnings, diagnostic messages, etc.
    Together with events you can log a custom set of additional parameters, 
    which are event-specific, scope-specific, 
    or common for all events from an app.
    Logging can be done simultaneously to R console, 
    JavaScript console in a browser, to a file log, and to a database (MongoDB).
    Log data can be analyzed with the help of process-mining techniques from 
    bupaR package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kalimu/shinyEventLogger#readme}
  \item \url{https://kalimu.github.io}
  \item Report bugs at \url{https://github.com/kalimu/shinyEventLogger/issues}
}

}
\author{
Kamil Wais \email{kamil.wais@gmail.com}
}
